Allow DM moving of client screen
