+++ client/src/game/operations/movement.ts
@@ -1,5 +1,6 @@
 import { addP, toArrayP, Vector } from "../../core/geometry";
 import { sendShapePositionUpdate } from "../api/emits/shape/core";
+import { moveClient } from "../client";
 import { Shape } from "../shapes/shape";
 import { TriangulationTarget, visionState } from "../vision/state";
 
@@ -50,7 +51,7 @@ export function moveShapes(shapes: readonly Shape[], delta: Vector, temporary: b
         // if (sel.refPoint.x % gridSize !== 0 || sel.refPoint.y % gridSize !== 0) sel.snapToGrid();
         if (!shape.preventSync) updateList.push(shape);
         if (shape.options.isPlayerRect ?? false) {
-            const x = 2;
+            moveClient(shape.uuid);
         }
     }
